+++ test/utils/exec.js
@@ -30,7 +30,20 @@ class TestContext {
 
   ledg(...args) {
     this._process = cp.spawnSync(this.ledg_bin, this.args.concat(args), this._modExecOpts({}));
+
+    if (this._process.stderr.length > 0 && this._debug) {
+      console.error(`Command: ${this.ledg_bin} ${this.args.concat(args).join(' ')}`);
+      console.error(`Stdin: ${this._input.toString()}`);
+      console.error(`Stdout: ${this._process.stdout.toString()}`);
+    }
+
     assert(this._process.stderr.length == 0, "Exited with stderr: " + this._process.stderr);
+
+    return this;
+  }
+
+  toggleDebug() {
+    this._debug = !this._debug;
     return this;
   }